+++ src/installer/SetupWizard/App.xaml.cs
@@ -1,0 +1,8 @@
+using System.Windows;
+
+namespace TycoonSetupWizard
+{
+    public partial class App : Application
+    {
+    }
+}